:recycle: Move types in pas.go to ast sub package
