+++ ast/type.go
@@ -1,0 +1,8 @@
+package ast
+
+func (TypeBlock) isFileSectionBlock() {}
+
+type TypeBlock []TypeDeclaration
+type TypeDeclaration interface {
+	isTypeDeclaration()
+}
