+++ ast/var.go
@@ -1,0 +1,12 @@
+package ast
+
+func (VarBlock) isFileSectionBlock() {}
+
+type VarBlock []*Variable
+
+func (*Variable) isClassMember() {}
+
+type Variable struct {
+	Name string
+	Type string
+}
